Convert from value types to classes.
Value types were starting to add too much complexity due to the need to inter-relate distinct hardware components together; e.g. memory being aware of the LCD Controller, which in turn needed to be aware of its own internal memory. All of these relationships modeled as value types required that all mutations propagate all the way up to the Gameboy instance, otherwise the hardware could get into odd / inconsistent states.

Instead, all of the hardware elements are now classes, and the LR35902 CPU has an internal "state" value type that can be used to take a snapshot of the CPU for testing purposes.